+++ Scripts/Runtime/Interfaces/ISymbolRule.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace PUCPR.AutoDefineSymbols
+{
+    public interface ISymbolRule
+    {
+        public ConditionalSymbolData[] GetSymbolAndConditionals(string[] refs);
+    }
+}