+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## [0.0.1] - YYYY-MM-DD
+## [1.0.0] - 2025-06-30
+
+### Added
+- Auto configurable dynamic Define Symbols by Rules
+- Rule for Package Define Symbols
+- Config file to center Rules
+
+
+## [0.0.1] - 2025-06-26
 
 ### Added
 - This CHANGELOG